+++ backend/src/main/java/styp/com/cenate/dto/PersonalCntRequest.java
@@ -72,5 +72,5 @@ public class PersonalCntRequest {
     
     private Long idArea;
     
-    private Integer idUsuario;
+    private Long idUsuario;  // ✅ Cambiado a Long
 }
